+++ pkg/stats/damage/damage.go
@@ -42,7 +42,7 @@ func NewStat(core *core.Core) (stats.Collector, error) {
 			return false
 		}
 
-		bucket := int(core.F / bucketSize)
+		bucket := core.F / bucketSize
 		last := out.cumu[len(out.cumu)-1]
 		for bucket >= len(out.cumu) {
 			newBucket := make([]float64, len(core.Player.Chars()))
